+++ statstracker.h
@@ -152,8 +152,8 @@ class ProductionStats
     vector< vector< vector<int> > > timeSpentGatheringWithDeviceByGroup;         // *** indexed by device_name_t, then group id, then day
     
     //BRH Two new stats for tracking time by resource
-    vector< vector< vector<double> > > timeSpentGatheringWithoutDeviceByAgentByRes;        //[resId][aId][day]
-    vector< vector< vector< vector<double> > > > timeSpentGatheringWithDeviceByAgentByRes; //[type][resId][aId][day]
+    vector< vector< vector<int> > > timeSpentGatheringWithoutDeviceByAgentByRes;        //[resId][aId][day]
+    vector< vector< vector< vector<int> > > > timeSpentGatheringWithDeviceByAgentByRes; //[type][resId][aId][day]
     /**
      * Time the society spent gathering devices each day.
      * indexed by device_name_t, then day
@@ -201,10 +201,10 @@ public:
 	vector< vector< vector<int> > > getDevicesMadeByRes() { return devicesMadeByRes; };     /*BRH 10.11.2017 */
     vector< vector< vector<int> > > getDevicesMadeByGroup() { return devicesMadeByGroup; };
     vector< vector<int> > getTimeSpentGatheringWithDevice() { return timeSpentGatheringWithDevice; };
-    vector< vector<vector< vector<double> > > > getTimeSpentGatheringWithDeviceByRes() { return timeSpentGatheringWithDeviceByAgentByRes; };
+    vector< vector<vector< vector<int> > > > getTimeSpentGatheringWithDeviceByRes() { return timeSpentGatheringWithDeviceByAgentByRes; };
     vector< vector<vector<int> > > getTimeSpentGatheringWithDeviceByGroup() { return timeSpentGatheringWithDeviceByGroup; };
     vector< int > getTimeSpentGatheringWithoutDevice() { return timeSpentGatheringWithoutDevice; };
-    vector< vector<vector<double> > > getTimeSpentGatheringWithoutDeviceByRes() { return timeSpentGatheringWithoutDeviceByAgentByRes; };
+    vector< vector<vector<int> > > getTimeSpentGatheringWithoutDeviceByRes() { return timeSpentGatheringWithoutDeviceByAgentByRes; };
     vector<vector<int> > getTimeSpentGatheringWithoutDeviceByGroup() { return timeSpentGatheringWithoutDeviceByGroup; };
     vector< vector<int> > getTimeSpentMakingDevices() { return timeSpentMakingDevices; };
     vector<vector<vector<int> > > getTimeSpentMakingDevicesByGroup() { return timeSpentMakingDevicesByGroup; };
